Moved mentors page to correct path
